Add initial support for Metal Performance Shaders backend.
It works generally well on Apple Silicon. However there are pytorch
operators, such as 'aten::_weight_norm_interface', which are not
currently implemented for the MPS device. As a temporary fix, you can
set the environment variable `PYTORCH_ENABLE_MPS_FALLBACK=1` to use
the CPU as a fallback for this op.

Similarly, when the XTTS engine is selected, MPS does not compose
nicely with deepspeed, because features such as redirects (i.e.,
torch.distributed.elastic.multiprocessing.redirects) are not
implemented on the CPU backend for Windows and MacOs at the time
of writing. Run alongside --no-deepspeed option at the command line,
where needed.

Everything else stays the same.

Committer: Alfonso De Gregorio <adg@secYOUre.com>